# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ForensiX Digital Forensics Investigation Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ActivityItem {
  id: number;
  description: string;
  timestamp: string;
  type: string;
}

export interface TypeCount {
  type: string;
  count: number;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface DashboardStats {
  totalCases: number;
  activeCases: number;
  totalEvidence: number;
  analyzedEvidence: number;
  recentActivity: ActivityItem[];
  evidenceByType: TypeCount[];
  casesByStatus: StatusCount[];
}

export type CaseStatus = typeof CaseStatus[keyof typeof CaseStatus];


export const CaseStatus = {
  open: 'open',
  active: 'active',
  closed: 'closed',
  archived: 'archived',
} as const;

/**
 * @nullable
 */
export type CasePriority = typeof CasePriority[keyof typeof CasePriority] | null;


export const CasePriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface Case {
  id: number;
  caseNumber: string;
  name: string;
  /** @nullable */
  description?: string | null;
  status: CaseStatus;
  /** @nullable */
  priority?: CasePriority;
  investigator: string;
  /** @nullable */
  tags?: string | null;
  createdAt: string;
  updatedAt: string;
  evidenceCount: number;
}

export type CaseInputPriority = typeof CaseInputPriority[keyof typeof CaseInputPriority];


export const CaseInputPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface CaseInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  /** @minLength 1 */
  investigator: string;
  priority?: CaseInputPriority;
  tags?: string;
}

export type CaseUpdateStatus = typeof CaseUpdateStatus[keyof typeof CaseUpdateStatus];


export const CaseUpdateStatus = {
  open: 'open',
  active: 'active',
  closed: 'closed',
  archived: 'archived',
} as const;

export type CaseUpdatePriority = typeof CaseUpdatePriority[keyof typeof CaseUpdatePriority];


export const CaseUpdatePriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface CaseUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  status?: CaseUpdateStatus;
  investigator?: string;
  priority?: CaseUpdatePriority;
  tags?: string;
}

export type EvidenceAnalysisStatus = typeof EvidenceAnalysisStatus[keyof typeof EvidenceAnalysisStatus];


export const EvidenceAnalysisStatus = {
  pending: 'pending',
  analyzing: 'analyzing',
  complete: 'complete',
  failed: 'failed',
} as const;

export interface Evidence {
  id: number;
  caseId: number;
  originalName: string;
  storedName: string;
  mimeType: string;
  fileSize: number;
  sha256: string;
  md5: string;
  sha1: string;
  analysisStatus: EvidenceAnalysisStatus;
  /** @nullable */
  notes?: string | null;
  uploadedAt: string;
}

export interface FileMetadata {
  name?: string;
  size?: number;
  mimeType?: string;
  /** @nullable */
  created?: string | null;
  /** @nullable */
  modified?: string | null;
  /** @nullable */
  accessed?: string | null;
  sha256?: string;
  md5?: string;
  sha1?: string;
}

export interface ExtractedStrings {
  urls?: string[];
  ipAddresses?: string[];
  emailAddresses?: string[];
  domains?: string[];
  printableStrings?: string[];
}

export interface DocumentMeta {
  /** @nullable */
  author?: string | null;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  subject?: string | null;
  /** @nullable */
  creator?: string | null;
  /** @nullable */
  producer?: string | null;
  /** @nullable */
  creationDate?: string | null;
  /** @nullable */
  modDate?: string | null;
  /** @nullable */
  pageCount?: number | null;
  /** @nullable */
  wordCount?: number | null;
}

export interface ImageMeta {
  /** @nullable */
  make?: string | null;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  software?: string | null;
  /** @nullable */
  dateTime?: string | null;
  /** @nullable */
  gpsLatitude?: number | null;
  /** @nullable */
  gpsLongitude?: number | null;
  /** @nullable */
  width?: number | null;
  /** @nullable */
  height?: number | null;
  /** @nullable */
  colorSpace?: string | null;
}

export interface EmailMeta {
  /** @nullable */
  from?: string | null;
  /** @nullable */
  to?: string | null;
  /** @nullable */
  subject?: string | null;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  messageId?: string | null;
  receivedChain?: string[];
  hasAttachments?: boolean;
  attachmentNames?: string[];
}

export type FindingSeverity = typeof FindingSeverity[keyof typeof FindingSeverity];


export const FindingSeverity = {
  info: 'info',
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface Finding {
  id: number;
  evidenceId: number;
  caseId?: number;
  severity: FindingSeverity;
  description: string;
  indicatorType: string;
  value: string;
  createdAt: string;
}

export interface AnalysisResult {
  evidenceId: number;
  status: string;
  fileMetadata?: FileMetadata;
  strings?: ExtractedStrings;
  documentMeta?: DocumentMeta;
  imageMeta?: ImageMeta;
  emailMeta?: EmailMeta;
  findings?: Finding[];
}

export interface TimelineEvent {
  id: number;
  caseId: number;
  /** @nullable */
  evidenceId?: number | null;
  /** @nullable */
  evidenceName?: string | null;
  timestamp: string;
  eventType: string;
  description: string;
  source: string;
}

export interface CustodyRecord {
  id: number;
  evidenceId: number;
  action: string;
  actor: string;
  /** @nullable */
  notes?: string | null;
  timestamp: string;
}

export type ReportInputFormat = typeof ReportInputFormat[keyof typeof ReportInputFormat];


export const ReportInputFormat = {
  json: 'json',
  html: 'html',
} as const;

export interface ReportInput {
  format: ReportInputFormat;
  title?: string;
  includeTimeline?: boolean;
  includeFindings?: boolean;
  includeHashes?: boolean;
  includeCustody?: boolean;
}

export interface Report {
  caseId: number;
  format: string;
  generatedAt: string;
  title?: string;
  content: string;
}

export interface YaraScanInput {
  evidenceId: number;
  /** Custom YARA-like pattern rules (one per line, regex patterns) */
  rules?: string;
  useBuiltinRules?: boolean;
}

export type YaraMatchSeverity = typeof YaraMatchSeverity[keyof typeof YaraMatchSeverity];


export const YaraMatchSeverity = {
  info: 'info',
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface YaraMatch {
  ruleName: string;
  description: string;
  matchedStrings: string[];
  severity: YaraMatchSeverity;
}

export interface YaraScanResult {
  evidenceId: number;
  matches: YaraMatch[];
  scannedAt: string;
  rulesUsed?: number;
}

export interface SearchResults {
  query: string;
  totalResults: number;
  cases: Case[];
  evidence: Evidence[];
  findings: Finding[];
}

export type SearchForensicParams = {
q: string;
type?: SearchForensicType;
};

export type SearchForensicType = typeof SearchForensicType[keyof typeof SearchForensicType];


export const SearchForensicType = {
  all: 'all',
  cases: 'cases',
  evidence: 'evidence',
  findings: 'findings',
} as const;