# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ForensiX Digital Forensics Investigation Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cases/{caseId}/timeline:
    get:
      operationId: getCaseTimeline
      tags: [timeline]
      summary: Get forensic timeline for a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Timeline events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelineEvent"
  /cases/{caseId}/report:
    post:
      operationId: generateReport
      tags: [reports]
      summary: Generate a forensic report for a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "200":
          description: Report generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Case not found
...
components:
  schemas:
...
    TimelineEvent:
      type: object
      required: [id, caseId, timestamp, eventType, description, source]
      properties:
        id:
          type: integer
        caseId:
          type: integer
        evidenceId:
          type: ["integer", "null"]
        evidenceName:
          type: ["string", "null"]
        timestamp:
          type: string
        eventType:
          type: string
        description:
          type: string
        source:
          type: string
...
    ReportInput:
      type: object
      required: [format]
      properties:
        format:
          type: string
          enum: [json, html]
        title:
          type: string
        includeTimeline:
          type: boolean
        includeFindings:
          type: boolean
        includeHashes:
          type: boolean
        includeCustody:
          type: boolean
    Report:
      type: object
      required: [caseId, format, generatedAt, content]
      properties:
        caseId:
          type: integer
        format:
          type: string
        generatedAt:
          type: string
        title:
          type: string
        content:
          type: string
...