# lib/api-zod/src/generated/types/reportInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ForensiX Digital Forensics Investigation Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ReportInputFormat } from './reportInputFormat';

export interface ReportInput {
  format: ReportInputFormat;
  title?: string;
  includeTimeline?: boolean;
  includeFindings?: boolean;
  includeHashes?: boolean;
  includeCustody?: boolean;
}