# lib/api-zod/src/generated/types/timelineEvent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ForensiX Digital Forensics Investigation Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface TimelineEvent {
  id: number;
  caseId: number;
  /** @nullable */
  evidenceId?: number | null;
  /** @nullable */
  evidenceName?: string | null;
  timestamp: string;
  eventType: string;
  description: string;
  source: string;
}